crypto-psa: Move the HMAC test to the crypto-psa lib
