+++ crypto/src/lib.rs
@@ -9,30 +9,3 @@ pub use edhoc_crypto_cc2538::*;
 #[cfg(feature = "psa")]
 pub use edhoc_crypto_psa::*;
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_hmac_sha256() {
-        const KEY: [u8; 32] = [0x0b; 32];
-        const MESSAGE_1: [u8; 0] = [];
-        const RESULT_1_TV: [u8; 32] = [
-            0x51, 0x77, 0xe6, 0x37, 0xaa, 0xac, 0x0b, 0x50, 0xe5, 0xdc, 0xa8, 0xbb, 0x05, 0xb0,
-            0xb5, 0x71, 0x44, 0x4b, 0xd5, 0x9b, 0x9b, 0x0d, 0x83, 0x4d, 0x50, 0x68, 0x1a, 0xf2,
-            0x1f, 0xc1, 0x4b, 0x1e,
-        ];
-        const MESSAGE_2: [u8; 1] = [0x0a];
-        const RESULT_2_TV: [u8; 32] = [
-            0x30, 0x50, 0x86, 0x79, 0x39, 0x85, 0x02, 0xd9, 0xdd, 0x70, 0x7e, 0xff, 0x6c, 0x84,
-            0x08, 0x9d, 0x83, 0x12, 0xcc, 0xea, 0x25, 0x36, 0x4d, 0x9c, 0xb8, 0xb0, 0xbd, 0x94,
-            0xd0, 0xe6, 0x55, 0xa3,
-        ];
-
-        let result_1 = hmac_sha256(&MESSAGE_1, KEY).to_public_array();
-        assert_eq!(result_1, RESULT_1_TV);
-
-        let result_2 = hmac_sha256(&MESSAGE_2, KEY).to_public_array();
-        assert_eq!(result_2, RESULT_2_TV);
-    }
-}
